Adding make config jobs config diff
Creating the updates for the config via ncurses is fine, but separating
it back out into the arch vs common config files for the build was being
done by hand. This allows a user to run `make menuconfig` from inside
the target build then come back to the openwrt dir and run the necessary
jobs to parse the configs appropriately.