Update openwrt upstream to Jan 31 2020
Currently only the netgear 3700v2, 3800[cb] hardware is working with
this upstream openwrt version. There are several open issues around
the firmware for the ath10k and related devices that need to be
addressed. We were hoping to get some of these fixes to relieve the
errors I was seeing when testing however it still remains an issue.

The c2600 TPLink is not able to be stable to service requests. More
testing is need but it appears that its related to the Linux OOM and the
newer firmware in the ath10k driver. The c2600 will be marked as broken
for now until the upstream openwrt broken comes up with a fix.
Unfortunately, this will not be in time for SCaLE 18x. At this time its
not a show stopper since we only have about a dozen pieces of this
hardware.