Fixing PR comment trigger for gitlab flash
The initial trigger that was added did not work and there wasnt a good
way to test this since it needed to land in master. Instead of trying to
guess at the fixes needed I opted to create a separate test repo for
this trigger. Everything works as expected now and sets up the "/tux"
command via PR comments for future tests.

Its setup like a CLI command, example:

/tux openwrt flash <WORMHOLE-STRING>
 |     |       |           |
 |     |       |           --- Arg needed for flashing
 |     |       --- Specific action to take
 |     --- Subcommand for specific components of the repo
 --- /tux trigger word