autoflash removal of sudo subcommands
Inside the autoflash script we were calling sudo for various commands
that needed elevated permissions to interface with things like gpio
pins, network interfaces, etc. This was causing (best I could tell)
processes like dnsmasq to be unable to daemon. The spawn would execute
but there would be no process left behind. I believe this was due to
there job in CI being run with `sudo -s ./autoflash` and being nested in
the sudo calls. There seems to be differing behavior on this newer
version of the gitlab-runner: 14.1.0 The fix was the drop all
superfluous sudo calls and assume we are already elevated which I think
is a reasonable expectation considering the nature of the autoflash
script.