Fixups for autoflash for unused/confusing syntax
Remove arp full path is not needed.

Also previous has "\n" when should have had "\r" for tftp
portion. tftp code seemed to be working regardless but
should correct this so its whats expected.

The was a lonely "close" in the autoflash script that I think was left
over "just in case" for the tftp command. This is also not needed and
has been since cleaned up.