apinger-pop for rc.local
Dealing with a race condition that comes up when there is NOT a dhcpd
server on the network. This causes the apinger template to never
become populated and thus the wifi radios remain up.

Adding the apinger template to default once to 8.8.8.8 regardless of
being triggered from the dhcp client.

Additionally, there seems to be a second race condition that if the
wireless radio is brought down too early in the startup process it
doesnt actually power down the radios. It leaves them in a weird state
of radios are up, leds are on, but the wifi cli tool thinks there off.