ensure wifi up prior to repopulating apinger.conf
Observed a race condition during testing where udhcpc client correctly
populated apinger.conf via apinger-pop.sh but the delayed catch all for
apinger.conf in /etc/rc.local then triggered after its 30 sec delay.
Since it was defaulting to 8.8.8.8 (google) and was not able to reach
out to the internet it downed the wifi which is the desired behavior.
However, then the dhcp renew triggered which reconfigured apinger.conf
setting it back to a valid ip to check (the gateway) but the wifi was
down and remained down as apinger did not need to trigger on
any alarms.

I believe its valid to leave this edge-case handled rather than fixing
/etc/rc.local to just check if a gateway is set. Mainly due to the case
were a bad gateway was present and downed wifi remains down if a new
gateway is handed out via dhcp and apinger is restarted.

Additional note: apinger reload doesnt keep the same pid. It still
restarts the service and thus causes the alarm to lose state.