mimic inventory.py outside of ansible
Our existing inventory.py is a very impressive aggregator of facts and
config with the scale-network repo. It seems best to continue to
leverage this but I would like to not effect the existing inventory.py
script for ansible: ansible/inventory.py

Instead Id like to propose making a copy of this and modifying it to fit
the new requirements for configuration for different services (like kea)
and configuration that is done outside of ansible.

After copying ansible/inventory.py to facts/inventory.py Ive made the
following temporary modifications as I work through updates to the
configuration:

- ignore anything from bridge vlan type
- simplify dhcpv4 to single ranges
- simplify dhcpv6 ranges to start-end
- complete kea config for v4
- fix 0 bitmask for subnet v4