Rewriting massflash in shell
massflash was originally written into expect so that we could pass N
passwords to ssh's prompt. Now that we have ssh keys being used
primarily and ssh password prompts disabled, this feature is not longer
valuable.

Converting it to a simple shell script and making it much more
straightfoward for how to massflash our APs.