flipping the flow of config gen to support N archs
Previously we were diffing the configs based on the prefixes of the
option names. Anything with that included CONFIG_TARGET was target
specific, while everything else was common.

This no longer works since there are many packages that exist
specifically for a given arch and are not prefixed with CONFIG_TARGET.

So instead the idea is to do the following order of operations to
generate our configurations across N architectures:

1. Generate a common.config using the x86 architecture since it omits
   any specific architecture packages. All packages declared here should
   be hardware independent and common across any architecture in scales
   environment.
2. For new architectures generate a $TARGET-generic.config without
   including any common config.
3. Combine $TARGET-generic.config + common.config and tweak until it
   includes all desired configuration. Output this the
   $TARGET-generic.config. This will include all CONFIG_TARGET config and
   architecture specific packages.