openwrt: disable uhttpd service on start
Some pkgs from opkg we require but dont need their default service to
start. There isnt a clear way to do this at build time so setting it via
rc.local seems to be the most straight forward approach

uhttpd is an example of such a service where its need for
prometheus but not by itself.