nix: bootstrap.sh UEFI labels and formatting
Had an issue on the hypervisors where the UUID of the disks was changing
between boots. I thought that the point of UUIDs was so this didnt
happen... Anyway defaulting to using FAT disk labels so we dont have to
concern ourselves with the UUIDs going forward.

Additionally, include the changes to the bootstrap script for formatting
disks with FAT32 instead of vfat for better compatibility.