nix: kea timeout to deal with interface binding
Was having an issue running microvm with qemu and the kea dhcp4 service not binding to all
interfaces. It would be ok if I `restarted` the service but it wouldnt come up cleanly on
its own. Even after trying to add additional requirements/deps to the unit file nothing
seemed to work. The following workaround was the only thing that seems to resolve the issue:

```
"interfaces-config": {
        ...
        "service-sockets-max-retries": 5,
        "service-sockets-retry-wait-time": 5000
    },
```
Ref: https://gitlab.isc.org/isc-projects/kea/-/issues/2776

There are other related issues to this:
- https://gitlab.isc.org/isc-projects/kea/-/issues/1084
- https://gitlab.isc.org/isc-projects/kea/-/issues/3188

Whats also odd is this wasnt an issue when running runNixOSTest locally which also use qemu.