nix: hypervisor2 dont wait for network online target
Give some of the network configurations we might be need for the
hypervisor boxes it would be best to skip waiting for this check. Its
meant to confirm that the network is actually functional. But in
practice, when its enabled it typically just causes nixos-rebuild
to take a long time before timing out.