gha: limit nixos builds to master and manual
Running these on every PR or every push to a PR quickly gets us
rate limited by github. Its unclear if theres a way around this cache
rate limit:
https://github.com/DeterminateSystems/magic-nix-cache-action?tab=readme-ov-file#usage-notes

Moving this to only running on `master` and manually should do the trick
for now. If it persists we might have to move it to a scheduled job like
openwrt builds.