nix.flake.inputs.nixpkgs-2405: pin for parity
While removing flake-parts, this input got bumped and is changing the
closures for packages and nixos systems. One of the goals during the
migration from flake-parts was to not have the closures change,
effectively a no-op. This input is being pinned so closures can be
compared between this branch and master to verify closures have not
changed.