nix.checks.duplicates-facts: fix fish warnings/errors
`fish` was raising warnings and errors because it is running in the
sandbox. It tries to write files as if it had access to the $HOME
directory but it does not. Adding the `--no-config` flag tells it not to
read from a configuration file so it will not go hunting for one or try
to write files where it cannot.

The warnings/errors below are what `fish` was writing to the shell.
These do not show up with this fix.

> duplicates-facts> error: can not save history
> duplicates-facts> warning-path: Unable to locate data directory derived from $HOME: '/homeless-shelter/.local/share/fish'.
> duplicates-facts> warning-path: The error was 'Permission denied'.
> duplicates-facts> warning-path: Please set $HOME to a directory where you have write access.
> duplicates-facts> error: can not save universal variables or functions
> duplicates-facts> warning-path: Unable to locate config directory derived from $HOME: '/homeless-shelter/.config/fish'.
> duplicates-facts> warning-path: The error was 'Permission denied'.
> duplicates-facts> warning-path: Please set $HOME to a directory where you have write access.