Update configuration.nix with promtail
Promtail is a part of the monitoring stack because it is the processor for logs. 

Promtail will send logs to a Loki service which is what injests the logs. Both services are needed, but promtail goes in front of Loki.

The Grafana dashboard will then be able to have Loki presets loaded into modules.

This push is just a build that puts the promtail servicee on the local NixOS host.